requirements.txt
beautifulsoup4
pysbd
python-dotenv
youtube-transcript-api
streamlit
ibm-watsonx-ai
